refactor commands to own module
