+++ src/nile/commands/install.py
