+++ tests/commands/test_test.py
@@ -1,0 +1,4 @@
+"""Test the test command."""
+
+def test_test():
+    assert True
